flask/classification_api/requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
flask-restx==1.1.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.13
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
mecab-ko-dic-msvc==0.999
mecab-ko-msvc==0.999
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyngrok==6.0.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
